backend/requirements.txt
Django==5.0.1
Pillow>=11.0.0
reportlab==4.0.9
python-dateutil==2.8.2

# Django REST Framework
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1

# Optional but recommended
django-filter==23.5
